requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
Flask==3.1.0
urllib3==2.2.3
